Separated out fused implementation of convolutions for clarity
