+++ tensorflow/core/kernels/conv_ops.h
@@ -16,8 +16,9 @@ limitations under the License.
 #ifndef TENSORFLOW_KERNELS_CONV_OPS_H_
 #define TENSORFLOW_KERNELS_CONV_OPS_H_
 
-#include "third_party/eigen3/unsupported/Eigen/CXX11/Tensor"
+#include "tensorflow/core/framework/resource_mgr.h"
 #include "tensorflow/core/util/tensor_format.h"
+#include "third_party/eigen3/unsupported/Eigen/CXX11/Tensor"
 
 #if GOOGLE_CUDA
 #include "tensorflow/core/kernels/conv_ops_gpu.h"
@@ -36,6 +37,14 @@ class LaunchConv2DOp {
               const Tensor& input, const Tensor& filter, int row_stride,
               int col_stride, const Eigen::PaddingType& padding, Tensor* output,
               TensorFormat data_format);
+};
+
+// Used to keep track of persistent memory buffers used within the op.
+template <class T, size_t size>
+struct Im2ColBufferResource : public ResourceBase {
+  mutex mu;
+  T data[size];
+  string DebugString() { return "Im2ColBufferResource"; }
 };
 
 #ifdef GOOGLE_CUDA
